+++ src/components/SidebarNav/SidebarNav.tsx
@@ -6,7 +6,7 @@ import TrickerTitle from '@assets/TrickerTitle'
 import { NavbarItem } from '@components/NavbarItem/NavbarItem'
 import Icon from '@components/Icon/Icon'
 import TimeTrackingBadge from '@components/TimeTrackingBadge/TimeTrackingBadge'
-import { ProfileButton } from '@components/ProfileButton/ProfileButton'
+import { ProfilePicture } from '@components/ProfilePicture/ProfilePicture'
 import Body1 from '@utils/typography/body1/body1'
 import { NavLink } from 'react-router-dom'
 import { type TimeTracking, type User } from '@utils/types'
@@ -104,7 +104,7 @@ export const SidebarNav: React.FC<SidebarNavProps> = ({
           {timeTracking && <TimeTrackingBadge ticketId={timeTracking.id} />}
           <NavLink to={'/user/' + user?.id}>
             <div className="flex items-center p-2 gap-3 max-w-[224px]">
-              <ProfileButton
+              <ProfilePicture
                 className="min-w-10 min-h-10"
                 img="https://s3-alpha-sig.figma.com/img/4fe8/a23d/ddeece2a91e7cc5919fd149d572c6d1e?Expires=1708905600&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4&Signature=a33zUweOtCPNcY1RYMBSl7M0W3HvLrpSgGfHnnqS-~FBATDlE42BrkMOby65VNWC2eo3p7sknPz1zjtO3xZfNT4zZyke6ZRrYV1k2nllK6NJMDzTKFn~qe4R0xWUtyxxWtauAlAvqmDY7G2O417AE05nFyFXyLlo7zePBrxsCNWm9f3jD2W65zFwgLy8wzcy5ryT5OZPA5wxOXPXN-6-VngmrBmoZqg-SWVfgL-E6W3GkoLj4IvMi7LcJZ162JsXmP0o-mHJ4bRi9K04k3ACjyg7BT2f9fCLbGzy5Nddzk8p61tDl7OczzCY-K9bx0ju3uAbhMnWfpFU0vcp7gpOcw__"
               />
